fix: 캐릭터 로직 위치 수정
